+++ ggpo/gui/ui/ggpowindow_ui.py
@@ -2,7 +2,7 @@
 
 # Form implementation generated from reading ui file 'ggpo/gui/ui/ggpowindow.ui'
 #
-# Created: Sat Mar 22 18:35:35 2014
+# Created: Sun Mar 23 17:49:56 2014
 #      by: PyQt4 UI code generator 4.9.1
 #
 # WARNING! All changes made in this file will be lost!
@@ -155,6 +155,9 @@ class Ui_MainWindow(object):
         self.uiShowCountryFlagInChatAct = QtGui.QAction(MainWindow)
         self.uiShowCountryFlagInChatAct.setCheckable(True)
         self.uiShowCountryFlagInChatAct.setObjectName(_fromUtf8("uiShowCountryFlagInChatAct"))
+        self.uiDisableAutoAnnounceAct = QtGui.QAction(MainWindow)
+        self.uiDisableAutoAnnounceAct.setCheckable(True)
+        self.uiDisableAutoAnnounceAct.setObjectName(_fromUtf8("uiDisableAutoAnnounceAct"))
         self.menuAction.addAction(self.uiAwayAct)
         self.menuAction.addAction(self.uiFocusOnChatAct)
         self.menuAction.addAction(self.uiEmoticonAct)
@@ -184,6 +187,7 @@ class Ui_MainWindow(object):
         self.menuSetting.addSeparator()
         self.menuSetting.addAction(self.uiNotifyPlayerStateChangeAct)
         self.menuSetting.addAction(self.uiShowCountryFlagInChatAct)
+        self.menuSetting.addAction(self.uiDisableAutoAnnounceAct)
         self.menuSetting.addAction(self.uiDebugLogAct)
         self.menuAbout.addAction(self.uiSRKForumAct)
         self.menuAbout.addAction(self.uiSRKWikiAct)
@@ -263,5 +267,6 @@ class Ui_MainWindow(object):
         self.action0.setText(QtGui.QApplication.translate("MainWindow", "0", None, QtGui.QApplication.UnicodeUTF8))
         self.uiCustomEmoticonsAct.setText(QtGui.QApplication.translate("MainWindow", "Custom &Emoticons", None, QtGui.QApplication.UnicodeUTF8))
         self.uiShowCountryFlagInChatAct.setText(QtGui.QApplication.translate("MainWindow", "Show &country flag in chat", None, QtGui.QApplication.UnicodeUTF8))
+        self.uiDisableAutoAnnounceAct.setText(QtGui.QApplication.translate("MainWindow", "&Disable auto announce in unsupported room", None, QtGui.QApplication.UnicodeUTF8))
 
 from ggpo.gui.completionlineedit import CompletionLineEdit